Add files via upload
We use a comprehensive database of the Swift GRBs with measured positions and spectroscopic redshifts.  Here we use the Swift XRT enhanced positional information, if available for a given event, otherwise we use the Swift BAT position.

The selection results in a total of 1665 GRBs in the blue.csv file.  Of this sample, 427 GRBs possess precise spectroscopic redshifts, with 300 derived from afterglow observations (bluea.csv) and 127 from host galaxy measurements (blueh.csv).